create iconSelector util
