+++ mobile/src/screens/Attendee/About.js
@@ -11,7 +11,7 @@ import data from '@data/data.json';
 /**
  * Returns the About screen
  * 
- * @param {object} props.navigation React Navigation navigation obtaining allowing for traversal to different screens
+ * @param {object} props.navigation React Navigation navigation object allowing for traversal to different screens
  * 
  * This screen presents conference information provided by the conference organizer
  * Allows navigation to view sponsors
