+++ mobile/src/screens/User/Profile.js
@@ -12,7 +12,7 @@ import data from '@data/data.json';
 /**
  * Returns the Profile screen
  * 
- * @param {object} props.navigation React Navigation navigation obtaining allowing for traversal to different screens
+ * @param {object} props.navigation React Navigation navigation object allowing for traversal to different screens
  * 
  * Displays the currently logged in user's contact information as well as the currently loaded conference instance
  */
